+++ src/Functions.Worker.Extensions.OpenAI/Assistants/ChatMessage.cs
@@ -1,7 +1,9 @@
 ﻿// Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-namespace Microsoft.Azure.WebJobs.Extensions.OpenAI.Models;
+using System.Text.Json.Serialization;
+
+namespace Microsoft.Azure.Functions.Worker.Extensions.OpenAI.Assistants;
 
 /// <summary>
 /// Chat Message Entity which contains the content of the message, the role of the chat agent, and the name of the calling function if applicable.
@@ -23,15 +25,18 @@ public class ChatMessage
     /// <summary>
     /// Gets or sets the content of the message.
     /// </summary>
+    [JsonPropertyName("content")]
     public string Content { get; set; }
 
     /// <summary>
     /// Gets or sets the role of the chat agent.
     /// </summary>
+    [JsonPropertyName("role")]
     public string Role { get; set; }
 
     /// <summary>
     /// Gets or sets the name of the calling function if applicable.
     /// </summary>
+    [JsonPropertyName("name")]
     public string? Name { get; set; }
 }